#747 dm-rest-commons should not depend on the DM
