+++ core/roboconf-dm-rest-commons/pom.xml
@@ -49,13 +49,6 @@
 		</dependency>
 		
 		<dependency>
-  			<groupId>net.roboconf</groupId>
-  			<artifactId>roboconf-dm</artifactId>
-  			<version>${project.version}</version>
-  			<scope>provided</scope>
-		</dependency>
-		
-		<dependency>
 			<groupId>com.sun.jersey</groupId>
 			<artifactId>jersey-core</artifactId>
 			<scope>provided</scope>
@@ -86,7 +79,14 @@
   			<version>${project.version}</version>
   			<type>test-jar</type>
 			<scope>test</scope>
-		</dependency>		
+		</dependency>
+		
+		<dependency>
+  			<groupId>net.roboconf</groupId>
+  			<artifactId>roboconf-dm</artifactId>
+  			<version>${project.version}</version>
+  			<scope>test</scope>
+		</dependency>	
 	</dependencies>
 	
 	<build>
