+++ core/roboconf-dm-rest-services/src/main/java/net/roboconf/dm/rest/services/internal/websocket/WebSocketHandler.java
@@ -40,8 +40,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import net.roboconf.core.model.beans.Application;
 import net.roboconf.core.model.beans.ApplicationTemplate;
 import net.roboconf.core.model.beans.Instance;
+import net.roboconf.core.model.runtime.EventType;
 import net.roboconf.core.utils.Utils;
-import net.roboconf.dm.management.events.EventType;
 import net.roboconf.dm.management.events.IDmListener;
 import net.roboconf.dm.rest.commons.beans.WebSocketMessage;
 import net.roboconf.dm.rest.commons.json.JSonBindingUtils;
