+++ core/roboconf-dm-scheduler/src/test/java/net/roboconf/dm/scheduler/internal/ManagerListenerTest.java
@@ -44,11 +44,11 @@ import org.quartz.impl.matchers.GroupMatcher;
 import net.roboconf.core.internal.tests.TestApplication;
 import net.roboconf.core.model.beans.Application;
 import net.roboconf.core.model.beans.ApplicationTemplate;
+import net.roboconf.core.model.runtime.EventType;
 import net.roboconf.core.utils.Utils;
 import net.roboconf.dm.internal.test.TestManagerWrapper;
 import net.roboconf.dm.management.ManagedApplication;
 import net.roboconf.dm.management.Manager;
-import net.roboconf.dm.management.events.EventType;
 
 /**
  * @author Vincent Zurczak - Linagora
