+++ core/roboconf-dm/src/main/java/net/roboconf/dm/internal/api/impl/NotificationMngrImpl.java
@@ -35,8 +35,8 @@ import java.util.logging.Logger;
 import net.roboconf.core.model.beans.Application;
 import net.roboconf.core.model.beans.ApplicationTemplate;
 import net.roboconf.core.model.beans.Instance;
+import net.roboconf.core.model.runtime.EventType;
 import net.roboconf.dm.management.api.INotificationMngr;
-import net.roboconf.dm.management.events.EventType;
 import net.roboconf.dm.management.events.IDmListener;
 
 /**
