+++ core/roboconf-dm/src/main/java/net/roboconf/dm/internal/utils/DmUtils.java
@@ -28,9 +28,9 @@ package net.roboconf.dm.internal.utils;
 import net.roboconf.core.model.beans.Instance;
 import net.roboconf.core.model.beans.Instance.InstanceStatus;
 import net.roboconf.core.model.helpers.InstanceHelpers;
+import net.roboconf.core.model.runtime.EventType;
 import net.roboconf.dm.management.ManagedApplication;
 import net.roboconf.dm.management.api.INotificationMngr;
-import net.roboconf.dm.management.events.EventType;
 
 /**
  * @author Vincent Zurczak - Linagora
