+++ core/roboconf-dm/src/main/java/net/roboconf/dm/management/events/IDmListener.java
@@ -28,6 +28,7 @@ package net.roboconf.dm.management.events;
 import net.roboconf.core.model.beans.Application;
 import net.roboconf.core.model.beans.ApplicationTemplate;
 import net.roboconf.core.model.beans.Instance;
+import net.roboconf.core.model.runtime.EventType;
 
 /**
  * An interface to propagate events generated by the DM.
