+++ karaf/roboconf-karaf-commands-dm/src/main/java/net/roboconf/karaf/commands/dm/diagnostics/DiagnosticMessagingCommand.java
@@ -43,8 +43,8 @@ import net.roboconf.core.model.beans.ApplicationTemplate;
 import net.roboconf.core.model.beans.Instance;
 import net.roboconf.core.model.beans.Instance.InstanceStatus;
 import net.roboconf.core.model.helpers.InstanceHelpers;
+import net.roboconf.core.model.runtime.EventType;
 import net.roboconf.dm.management.Manager;
-import net.roboconf.dm.management.events.EventType;
 import net.roboconf.dm.management.events.IDmListener;
 import net.roboconf.karaf.commands.dm.KarafDmCommandsUtils;
 import net.roboconf.karaf.commands.dm.KarafDmCommandsUtils.RbcfInfo;
